rename debugging -> cli
